feat: :bug:

+++ src/app/api/convert/route.ts
@@ -1,22 +1,34 @@
 import { ContentFormat, convertDocxToHtmlWithImage, convertDocxToHtmlWithouchImage } from '@/app/api/mammoth.utils';
 import { directory } from '@/core/provider/pathProvider';
 import fs from 'fs';
-import { readdir, unlink } from 'fs/promises';
+import { unlink } from 'fs/promises';
+import { StatusCodes } from 'http-status-codes';
 import { NextRequest, NextResponse } from 'next/server';
-import { join } from 'path';
-
+import path, { join } from 'path';
+import { checkDirectoryExists, checkFileExists, logger, readDirectory, sanitizeData } from '../utils/functions.utils';
 
 export async function GET() {
   try {
-    const targetFolder = join(process.cwd(), directory.HtmlPath);
-    const files = await readdir(targetFolder);
+    const folderExists = await checkDirectoryExists(directory.HtmlPath);
 
-    const data = files.filter(item => item !== '.gitkeep')
-
-    return NextResponse.json({ success: true, data });
+    if (folderExists) {
+      const files = await readDirectory(directory.HtmlPath);
+      const sanitizedData = sanitizeData(files);
+      return NextResponse.json({ success: true, data: sanitizedData }, { status: StatusCodes.OK });
+    }
   } catch (error) {
-    console.error(error);
-    return NextResponse.json({ success: false, error: "Failed to retrieve files" });
+    if (error instanceof Error) {
+      if (error.name === 'Error' && error.message.startsWith('ENOENT')) {
+        logger.error(`Directory not found: ${directory.HtmlPath}`);
+        return NextResponse.json({ success: false, error: "Directory not found" });
+      } else {
+        logger.error(error);
+        return NextResponse.json({ success: false }, {status: StatusCodes.BAD_REQUEST});
+      }
+    } else {
+      logger.error('Unknown error:', error);
+      return NextResponse.json({ success: false, error: "Unknown error" });
+    }
   }
 }
 
@@ -26,46 +38,54 @@ export async function POST(request: Request) {
     const { convertFormat } = requestBody
     const docxFiles = await fs.promises.readdir(directory.DocxPath);
 
-    if (!fs.existsSync(directory.HtmlPath)) {
-      fs.mkdirSync(directory.HtmlPath);
+    const directoryExists = await fs.promises.access(directory.HtmlPath)
+      .then(() => true)
+      .catch(() => false);
+
+    if (!directoryExists) {
+      await fs.promises.mkdir(directory.HtmlPath, { recursive: true });
     }
 
-    const convertedFiles = [];
-
-    for (const docxFile of docxFiles) {
-      if (docxFile.endsWith('.docx')) {
+    const convertedFiles = await Promise.all(docxFiles.map(async (docxFile) => {
+      if (path.extname(docxFile).toLowerCase() === '.docx') {
         const docxPath = join(directory.DocxPath, docxFile);
         let htmlContent = '';
 
-        if (convertFormat === ContentFormat.WithMetaAndImages) {
-          htmlContent = await convertDocxToHtmlWithImage(docxPath);
-        } else if (convertFormat === ContentFormat.WithoutMetaAndImages) {
-          htmlContent = await convertDocxToHtmlWithouchImage(docxPath);
+        switch (convertFormat) {
+          case ContentFormat.WithMetaAndImages:
+            htmlContent = await convertDocxToHtmlWithImage(docxPath);
+            break;
+          case ContentFormat.WithoutMetaAndImages:
+            htmlContent = await convertDocxToHtmlWithouchImage(docxPath);
+            break;
         }
 
         const htmlFileName = `${
           convertFormat === ContentFormat.WithMetaAndImages ? 'editor_' : ''
         }${docxFile.replace('.docx', '.html')}`;
 
-        const outputPath = join(directory.HtmlPath, htmlFileName);
-        await fs.promises.writeFile(outputPath, htmlContent);
+        const outputPath = path.resolve(directory.HtmlPath, htmlFileName);
 
-        convertedFiles.push({ html: htmlFileName });
+        const writeStream = fs.createWriteStream(outputPath);
+        writeStream.write(htmlContent);
+        writeStream.end();
+
+        return { html: htmlFileName };
       }
-    }
+    }));
 
     return new NextResponse(
       JSON.stringify({
         message: 'Conversion Complete',
         convertedFiles: convertedFiles,
       }),
-      { status: 200, headers: { 'Content-Type': 'application/json' } }
+      { status: StatusCodes.OK, headers: { 'Content-Type': 'application/json' } }
     );
   } catch (err) {
-    console.error('Error converting:', err);
+    logger.error('Error converting:', err);
     return new NextResponse(
       JSON.stringify({ error: 'Conversion failed', err }),
-      { status: 500, headers: { 'Content-Type': 'application/json' } }
+      { status: StatusCodes.BAD_REQUEST, headers: { 'Content-Type': 'application/json' } }
     );
   }
 }
@@ -74,11 +94,16 @@ export const DELETE = async (request: NextRequest) => {
   try {
     const header = request.headers.get('x-invoke-query');
 
-    if (!header) return NextResponse.json('Arquivo não encontrado ou inexistente', { status: 404 })
+    if (!header) return NextResponse.json('Arquivo não encontrado ou inexistente', { status: StatusCodes.NOT_FOUND })
 
-    const decodedString = decodeURIComponent(header);
-
-    const parsedObject = JSON.parse(decodedString);
+    let parsedObject;
+    try {
+      const decodedString = decodeURIComponent(header);
+      parsedObject = JSON.parse(decodedString);
+    } catch (error) {
+      logger.error('Error parsing JSON:', error);
+      return NextResponse.json({ message: 'Error parsing JSON' }, { status: StatusCodes.BAD_REQUEST });
+    }
 
     const fileName = parsedObject.filename;
 
@@ -86,14 +111,16 @@ export const DELETE = async (request: NextRequest) => {
       return NextResponse.json({ success: false, message: 'Filename not provided' });
     }
 
-    const targetFolder = join(process.cwd(), directory.HtmlPath);
-    const filePath = join(targetFolder, fileName);
+    const filePath = join(directory.HtmlPath, fileName);
+    const fileExists = await checkFileExists(filePath);
 
-    await unlink(filePath);
+    if (fileExists) {
+      await unlink(filePath);
+    }
 
-    return NextResponse.json({ success: true, message: 'Arquivo Deletado com Sucesso' }, { status: 200 });
+    return NextResponse.json({ success: true, message: 'Arquivo Deletado com Sucesso' }, { status: StatusCodes.OK });
   } catch (error) {
-    console.error('Error deleting file:', error);
-    return NextResponse.json({ message: 'Error deleting file' }, { status: 400 });
+    logger.error('Error deleting file:', error);
+    return NextResponse.json({ message: 'Error deleting file' }, { status: StatusCodes.BAD_REQUEST });
   }
 }
